Code review comment fixes continued.
